+++ sdk/core/core-client/test/serviceClient.spec.ts
@@ -24,6 +24,7 @@ import {
   HttpClient,
   createPipelineRequest,
   PipelineRequest,
+  RestError,
 } from "@azure/core-rest-pipeline";
 
 import {
@@ -256,7 +257,7 @@ describe("ServiceClient", function () {
     assert.deepEqual(request!.headers.toJSON(), expected);
   });
 
-  it("should call rawResponseCallback with the full response", async function () {
+  it("should call onResponse with the full response", async function () {
     let request: OperationRequest;
     const client = new ServiceClient({
       httpClient: {
@@ -292,6 +293,66 @@ describe("ServiceClient", function () {
     assert.strictEqual(rawResponse?.status, 200);
     assert.strictEqual(rawResponse?.request, request!);
     assert.strictEqual(rawResponse?.headers.get("X-Extra-Info"), "foo");
+  });
+
+  it("should call onResponse with the full response when encountering an unknown status", async function () {
+    let request: OperationRequest;
+
+    const pipeline = createEmptyPipeline();
+    pipeline.addPolicy(deserializationPolicy());
+    const client = new ServiceClient({
+      httpClient: {
+        sendRequest: (req) => {
+          request = req;
+          return Promise.resolve({
+            request,
+            status: 500,
+            headers: createHttpHeaders(),
+          });
+        },
+      },
+      pipeline,
+    });
+
+    let rawResponse: FullOperationResponse | undefined;
+    let requestFailed = false;
+    let caughtError: RestError | undefined;
+    let flatResponse: any;
+    let onResponseError: unknown;
+
+    try {
+      await client.sendOperationRequest(
+        {
+          options: {
+            onResponse: (response, flat, error) => {
+              rawResponse = response;
+              flatResponse = flat;
+              onResponseError = error;
+            },
+          },
+        },
+        {
+          httpMethod: "GET",
+          baseUrl: "https://example.com",
+          serializer: createSerializer(),
+          headerParameters: [],
+          responses: {
+            200: {},
+          },
+        }
+      );
+    } catch (e: any) {
+      caughtError = e;
+      requestFailed = true;
+      assert.strictEqual(e.name, "RestError");
+    }
+
+    assert(requestFailed, "Request should fail with unknown status");
+    assert(request!);
+    assert.strictEqual(rawResponse?.status, 500);
+    assert.strictEqual(rawResponse?.request, request!);
+    assert.deepStrictEqual(flatResponse, { body: undefined });
+    assert.strictEqual(caughtError, onResponseError);
   });
 
   it("should serialize collection:csv query parameters", async function () {